refactor(docs): extract CLI reference into lib/cli-reference.ts
Second cut of the docs generator split. Moves the CLI reference renderer
cluster into lib/cli-reference.ts: the CliCommand type, the cliQuickLinks +
cliReferenceGroups data tables, parseCliUsage / displayCliPartToken /
describeCliPart (usage-token parsing), renderCliCommandBlock, and
renderCliReference. Pure over CLI_COMMANDS (imported from core) + the
doc-utils escapers; the orchestrator imports renderCliReference back.
CLI_COMMANDS stays imported in generate.ts (8 other uses — snippet, coverage).

Pure relocation — bodies byte-for-byte unchanged. generate.ts 2,308 -> 2,022.
Verified: eslint (changed files) clean + docs:generate output BYTE-IDENTICAL
to baseline.
